requirements.txt
python-binance
pyyaml